requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
idna==2.10
install==1.3.4
matrix-client==0.3.2
python-dotenv==0.17.1
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.5
